* Configuring SI ConfigurableCompositeMessageConverter  - Moved ContentType related beans into separate configuration  - Configured SI ConfigurableCompositeMessageConverter to use same    converters as Stream does (for ServiceActivator) - TupleConverter should return byte[] as all other converters - Fixed tests
