+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/broken/InboundJsonToTupleConversionTest.java
@@ -56,12 +56,13 @@ public class InboundJsonToTupleConversionTest {
 	@Test
 	public void testInboundJsonTupleConversion() throws Exception {
 		testProcessor.input().send(MessageBuilder.withPayload("{'name':'foo'}")
-				.setHeader(MessageHeaders.CONTENT_TYPE, "application/json").build());
+				.build());
 		@SuppressWarnings("unchecked")
 		Message<?> received = ((TestSupportBinder) binderFactory.getBinder(null, MessageChannel.class))
 				.messageCollector().forChannel(testProcessor.output()).poll(1, TimeUnit.SECONDS);
 		assertThat(received).isNotNull();
-		assertThat(received.getPayload()).isEqualTo(TupleBuilder.tuple().of("name", "foo"));
+
+		assertThat(TupleBuilder.fromString(new String((byte[])received.getPayload()))).isEqualTo(TupleBuilder.tuple().of("name", "foo"));
 	}
 
 	@EnableBinding(Processor.class)
